requirements.txt
pyyaml
dictor
pyroute2
xmltodict
psutil